requirements.txt
alembic==1.13.2
fastapi==0.111.0
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
passlib==1.7.4
colorlog==6.8.2
pydantic==2.7.4
python-jose==3.3.0
uvicorn~=0.30.1
python-dotenv~=1.0.1